requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.7.2
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
billiard==4.2.0
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
Django==4.1.13
django-cors-headers==4.4.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
docopt==0.6.2
drf-yasg==1.21.7
einops==0.8.0
ffmpeg-python==0.2.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.5
HyperPyYAML==1.2.2
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.5
kombu==5.3.4
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.5
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy<2.0.0
omegaconf==2.3.0
openai==1.35.13
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
prompt-toolkit==3.0.43
protobuf==5.27.2
psycopg2-binary==2.9.9
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyAudioAnalysis==0.3.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==3.12.3
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytorch-lightning==2.3.3
pytorch-metric-learning==2.5.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.31
sqlparse==0.2.4
sympy==1.13.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.12
whisper==1.1.10
yarl==1.9.4
